pyproject.toml
[project]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Martynas09",email = "martynas200012@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest-playwright (>=0.7.0,<0.8.0)",
]

[tool.poetry]
packages = [{include = "poetry_demo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
